Clinic system/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
# psycopg2-binary==2.9.9  # Commented out - requires PostgreSQL dev libraries
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Monitoring & Logging
structlog==23.2.0

# CORS
python-cors==1.0.1

# Environment
python-dotenv==1.0.0
